add Player class, player object and input function
Added a Player class that handles the logic and rendering of the player
character.

Added an instance of that class.

Added a function ('process_input' for now) that runs the pygame input
logic. Inside it is a player instance method that takes care of the
player related input.

Finally, an assets' directory was added. For now, it contains some
images of the player.